# contracts/console-api.openapi.yaml
openapi: 3.0.3
info:
  title: Keel 控制台 API
  version: 1.0.0-draft
  description: |
    keel-console 九个页面消费的全部接口，由 keel-server 实现。

    这份文件是前端、mock 数据、后端实现三方的唯一事实来源：
    - 前端只调这里定义的接口，不直接调 Langfuse / 薄网关 / Prometheus
    - mock 数据由这份文件生成，接口切真时前端只改 baseURL
    - 后端 P1 按这份文件实现，字段不一致以这里为准

    和 keel/v1 五份契约的关系：那五份是智能体侧协议（/v1/invoke、审计事件、
    追踪属性、manifest、评测集），这份是控制面 REST API，两者独立演进。
    引用到的枚举值必须和 contracts/ 下对应 schema 保持一致。
servers:
  - url: https://keel.example.com/api/v1
    description: prod
  - url: http://localhost:5173/api/v1
    description: 本地 mock（MSW 拦截）
security:
  - bearerAuth: []
...
paths:
...
  /agents:
    get:
      tags: [agents]
      operationId: listAgents
      summary: 智能体列表
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: category
          in: query
          schema: { type: string, enum: [all, biz, dev], default: all }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/AgentStatus" }
        - name: q
          in: query
          description: 按 id、名称、负责人模糊匹配
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/AgentSummary" }
    post:
      tags: [agents]
      operationId: registerAgent
      summary: 注册智能体（keel register / 新建向导第四步）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: 注册完成，返回自检报告
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/SelfCheckReport" }
        "409":
          $ref: "#/components/responses/Error"
  /agents/{name}:
    parameters:
      - $ref: "#/components/parameters/AgentName"
    get:
      tags: [agents]
      operationId: getAgent
      summary: 智能体详情（抽屉）
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/AgentDetail" }
        "404":
          $ref: "#/components/responses/Error"
  /agents/{name}/chat:
    parameters:
      - $ref: "#/components/parameters/AgentName"
    post:
      tags: [agents]
      operationId: chatWithAgent
      summary: 向已注册的智能体发一轮对话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: 智能体的最终回复
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          text: { type: string }
                          traceId: { type: string, nullable: true }
        "404":
          $ref: "#/components/responses/Error"
...
  /agents/name-check:
    get:
      tags: [agents]
      operationId: checkAgentName
      summary: 新建向导第二步：名称可用性校验
      parameters:
        - name: name
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          available: { type: boolean }
                          reason:
                            type: string
                            nullable: true
                            description: 不可用时的原因，如「已被占用」「必须小写字母开头」
  /agents/manifest-preview:
    post:
      tags: [agents]
      operationId: previewManifest
      summary: 新建向导第四步：由表单生成 agent.yaml 预览
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          yaml: { type: string, description: agent.yaml 全文 }
                          warnings:
                            type: array
                            items: { type: string }
...
  /tools:
    get:
      tags: [tools]
      operationId: listTools
      summary: 工具注册表
      parameters:
        - $ref: "#/components/parameters/Env"
        - name: scope
          in: query
          schema: { type: string, enum: [all, PRIVATE, SHARED], default: all }
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ToolStatus" }
        - name: risk
          in: query
          schema: { $ref: "#/components/schemas/Risk" }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: "#/components/schemas/PageMeta"
                          - type: object
                            properties:
                              items:
                                type: array
                                items: { $ref: "#/components/schemas/ToolSummary" }
    post:
      tags: [tools]
      operationId: registerTool
      summary: 注册 MCP 工具
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ToolRegisterRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /tools/{name}:
    parameters:
      - $ref: "#/components/parameters/ToolName"
    get:
      tags: [tools]
      operationId: getTool
      summary: 工具详情（抽屉）：策略、依赖方、版本历史
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/ToolDetail" }
  /tools/{name}/versions:
    parameters:
      - $ref: "#/components/parameters/ToolName"
    post:
      tags: [tools]
      operationId: publishToolVersion
      summary: 发布新版本
      description: |
        描述或参数变化会触发所有依赖方回归评测。
        破坏兼容必须用新名字（如 work_order_create.v2），不能在原名上发版。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version, schemaJson, breaking]
              properties:
                version: { type: string }
                description: { type: string }
                schemaJson: { type: object, additionalProperties: true }
                access: { $ref: "#/components/schemas/ToolAccess" }
                risk: { $ref: "#/components/schemas/Risk" }
                breaking: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          triggeredRegressions:
                            type: array
                            description: 被触发回归评测的依赖方
                            items: { type: string }
        "400":
          description: breaking=true 时必须改名
          $ref: "#/components/responses/Error"
  /tools/{name}/deprecate:
    parameters:
      - $ref: "#/components/parameters/ToolName"
    post:
      tags: [tools]
      operationId: deprecateTool
      summary: 废弃（必须给替代工具和截止日期）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replacedBy, deadline]
              properties:
                replacedBy: { type: string }
                deadline: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
  /tools/{name}/retire:
    parameters:
      - $ref: "#/components/parameters/ToolName"
    post:
      tags: [tools]
      operationId: retireTool
      summary: 下线
      description: 有 prod 依赖方时返回 409，只能先废弃。
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Envelope" }
        "409":
          description: 存在 prod 依赖，拒绝下线
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: auth-gateway 签发，aud=keel-api
  parameters:
    Env:
      name: env
      in: query
      schema: { type: string, enum: [all, dev, test, staging, prod], default: all }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, enum: [10, 20, 50, 100], default: 10 }
    AgentName:
      name: name
      in: path
      required: true
      schema: { type: string }
    ToolName:
      name: name
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: 统一错误体
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Envelope:
      type: object
      required: [code, traceId]
      properties:
        code:
          type: string
          description: 成功固定为 OK，失败为 contracts/error-codes.yaml 里的错误码
          example: OK
        message: { type: string, nullable: true }
        traceId: { type: string }
        data: {}
    ApiError:
      type: object
      required: [code, message, traceId, retryable]
      properties:
        code: { type: string, example: TOOL_NOT_GRANTED }
        message: { type: string }
        traceId: { type: string }
        retryable: { type: boolean }
    PageMeta:
      type: object
      required: [page, size, total]
      properties:
        page: { type: integer }
        size: { type: integer }
        total: { type: integer }
    EnvName:
      type: string
      enum: [dev, test, staging, prod]
    Risk:
      type: string
      enum: [LOW, MID, HIGH]
    AgentStatus:
      type: string
      description: |
        DRAFT 表示 keel new 预占了名称但还没 register。
        原型里的「门禁未过」不是独立状态，由 status=REGISTERED 且 gatePassed=false 推导。
      enum: [DRAFT, REGISTERED, ONLINE, DEGRADED, OFFLINE, RETIRED]
...
    ToolStatus:
      type: string
      enum: [REGISTERED, ONLINE, DEPRECATED, RETIRED]
    ToolAccess:
      type: string
      enum: [READ, WRITE, EXEC]
    ToolScope:
      type: string
      enum: [PRIVATE, SHARED]
...
    Alert:
      type: object
      properties:
        at: { type: string, format: date-time }
        level: { type: string, enum: [INFO, WARN, ERROR] }
        agent: { type: string, nullable: true }
        kind:
          type: string
          description: 对账 finding 或运行告警
          enum:
            [
              OFFLINE,
              UNREGISTERED,
              ZOMBIE,
              RETIRE_INCOMPLETE,
              VERSION_MISMATCH,
              BUDGET_WARN,
              GATE_FAILED,
              PROMPT_DRIFT,
            ]
        text: { type: string }
        link: { type: string, nullable: true }

    # ------------------------------------------------------------ 智能体
    AgentSummary:
      type: object
      properties:
        name: { type: string, example: offshore-wind }
        displayName: { type: string, example: 风机维检 }
        category: { type: string, enum: [biz, dev] }
        language: { type: string, example: Python }
        runtime: { type: string, enum: [code, dify] }
        template: { type: string, example: graph-agent }
        env: { $ref: "#/components/schemas/EnvName" }
        version: { type: string, example: v0.7.2 }
        status: { $ref: "#/components/schemas/AgentStatus" }
        statusNote:
          type: string
          nullable: true
          example: 0/1 就绪 · CrashLoop
        gatePassed: { type: boolean, nullable: true }
        ownerOrg: { type: string }
        ownerUser: { type: string }
        calls24h: { type: integer }
        callsTotal: { type: integer }
        p95Seconds: { type: number, format: float }
        costCny: { type: number, format: float }
        dailyBudgetCny: { type: number, format: float }
        score: { type: number, format: float, nullable: true }
        instances: { type: string, example: "2/2", description: 就绪/期望；Dify 为探活结果 }
        multiAgent: { type: boolean }
        delegateCount:
          type: integer
          description: spec.delegates 的个数，列表卡片用来显示「编排 N 个」
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/AgentSummary"
        - type: object
          properties:
            manifestYaml: { type: string }
            manifestHash: { type: string }
            knowledgeBases:
              type: array
              items: { type: string }
            tools:
              type: array
              items: { type: string }
            models:
              type: array
              items: { type: string }
            delegates:
              type: array
              items: { type: string }
            selfCheck: { $ref: "#/components/schemas/SelfCheckReport" }
            instanceList:
              type: array
              items:
                type: object
                properties:
                  instanceId: { type: string }
                  source: { type: string, enum: [k8s, heartbeat, probe] }
                  version: { type: string }
                  ready: { type: boolean }
                  lastSeenAt: { type: string, format: date-time }
            resources:
              type: array
              description: 替它开通的外部资源，下线时按此回收
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [litellm_key, langfuse, secret, dataset, oauth_client]
                    description: litellm_key 是历史枚举值，语义为薄网关虚拟 Key
                  externalId: { type: string }
                  status: { type: string }
            versions:
              type: array
              items: { $ref: "#/components/schemas/ReleaseRecord" }
            findings:
              type: array
              description: 对账发现的问题
              items: { $ref: "#/components/schemas/Alert" }
    SelfCheckReport:
      type: object
      properties:
        passed: { type: boolean }
        items:
          type: array
          items:
            type: object
            properties:
              name: { type: string, example: 高风险工具绑定审批策略 }
              passed: { type: boolean }
              detail: { type: string, nullable: true }
    ReleaseRecord:
      type: object
      properties:
        version: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        gatePassed: { type: boolean }
        scoreTotal: { type: number, format: float, nullable: true }
        scoreByTag:
          type: object
          additionalProperties: { type: number, format: float }
        image: { type: string, nullable: true }
        ciRunUrl: { type: string, nullable: true }
        releasedBy: { type: string }
        releasedAt: { type: string, format: date-time }
    RegisterRequest:
      type: object
      required: [name, env]
      description: 向导表单；后端据此生成 agent.yaml。字段与 manifest.schema.json 一一对应。
      properties:
        name: { type: string }
        displayName: { type: string }
        category: { type: string, enum: [biz, dev] }
        ownerOrg: { type: string }
        ownerUser: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        runtime: { type: string, enum: [code, dify] }
        language: { type: string }
        template: { type: string }
        endpoint: { type: string }
        models:
          type: object
          properties:
            default: { type: string }
            fallback:
              type: array
              items: { type: string }
            dailyBudgetCny: { type: number, format: float }
            maxSteps: { type: integer }
        knowledgeBases:
          type: array
          items: { type: string }
        tools:
          type: array
          items: { type: string }
        delegates:
          type: array
          items: { type: string }
        approvalPolicyId: { type: string, nullable: true }
...
    ToolSummary:
      type: object
      properties:
        name: { type: string, example: work_order_create }
        description: { type: string }
        scope: { $ref: "#/components/schemas/ToolScope" }
        ownerAgent: { type: string, nullable: true }
        ownerOrg: { type: string, nullable: true }
        access: { $ref: "#/components/schemas/ToolAccess" }
        risk: { $ref: "#/components/schemas/Risk" }
        version: { type: string }
        status: { $ref: "#/components/schemas/ToolStatus" }
        approvalPolicy: { type: string, nullable: true, example: 班组长审批 }
        calls24h: { type: integer }
        dependentCount: { type: integer }
        replacedBy: { type: string, nullable: true }
        deprecateDeadline: { type: string, format: date, nullable: true }
    ToolDetail:
      allOf:
        - $ref: "#/components/schemas/ToolSummary"
        - type: object
          properties:
            provider: { type: string, nullable: true, description: MCP 服务地址 }
            schemaJson: { type: object, additionalProperties: true }
            dependents:
              type: array
              items: { $ref: "#/components/schemas/ToolDependent" }
            versions:
              type: array
              items:
                type: object
                properties:
                  version: { type: string }
                  change: { type: string }
                  breaking: { type: boolean }
                  at: { type: string, format: date-time }
    ToolDependent:
      type: object
      properties:
        agent: { type: string }
        env: { $ref: "#/components/schemas/EnvName" }
        agentStatus: { $ref: "#/components/schemas/AgentStatus" }
        declaredInManifest: { type: boolean }
        calls30d: { type: integer }
    ToolRegisterRequest:
      type: object
      required: [name, scope, access, risk, provider, schemaJson]
      properties:
        name: { type: string }
        description: { type: string }
        scope: { $ref: "#/components/schemas/ToolScope" }
        provider: { type: string, description: MCP 服务地址 }
        access: { $ref: "#/components/schemas/ToolAccess" }
        risk: { $ref: "#/components/schemas/Risk" }
        approvalPolicyId: { type: string, nullable: true }
        schemaJson: { type: object, additionalProperties: true }

    # ------------------------------------------------------------ 审批
...